requirements.txt
streamlit
google-generativeai>=0.5.0  # <--- 최신 버전을 받아오도록 수정
PyGithub
pandas